Revert unworkable structure
